Moved one_rank_only to idist, adapted tests
